scripts

+++ script/build.sh
@@ -1,24 +1,9 @@
 #!/bin/bash
 set -e
 
-# inizializza moduli
-source /etc/profile
-source /etc/profile.d/modules.sh
+command -v mpicc  >/dev/null 2>&1 || { echo "ERROR: mpicc not found. Run ./scripts/env.sh first."; exit 1; }
+command -v mpic++ >/dev/null 2>&1 || { echo "ERROR: mpic++ not found. Run ./scripts/env.sh first."; exit 1; }
 
-module purge
-module load gcc91
-module load openmpi-3.0.0--gcc-9.1.0
-module load cmake-3.15.4
-
-export OMPI_CC=/apps/gcc-9.1.0/local/bin/gcc-9.1.0
-export OMPI_CXX=/apps/gcc-9.1.0/local/bin/g++-9.1.0
-export OMPI_MCA_mpi_cuda_support=0
-export OMPI_MCA_oob=tcp
-export OMPI_MCA_btl="^openib"
-export OMPI_MCA_orte_base_help_aggregate=1
-
-
-# build (isolato)
 (
   cd "$(dirname "$0")/.."
   rm -rf build
@@ -34,10 +19,6 @@ export OMPI_MCA_orte_base_help_aggregate=1
 )
 
 echo
-echo "=== BUILD COMPLETATA ==="
-echo "Ambiente MPI/OMP caricato."
-echo "Ora puoi usare mpirun direttamente."
+echo "=== BUILD COMPLETED ==="
+echo "Now you can run the code"
 echo
-
-# sostituisce la shell corrente con una nuova che eredita l'ambiente
-exec bash
